test/integration: retire metrics-proxy self-test driver
The metrics-proxy self-test driver in test/integration/metrics_test.go
never runs in upstream CI. The nightly KIND cluster does not install
the prometheus-operator CRDs, so monitoring.coreos.com is absent. The
bpfman-operator gates its metrics-proxy DaemonSet reconcile on that
API group (controllers/bpfman-operator/config.go:328) and the test
guards itself with the same check, hitting t.Skip on every nightly
run.

Delete metrics_test.go and move its podExec helper to common.go,
which also uses podExec for link-ordering checks. The rest of the
test file -- ServiceAccount/token plumbing, pod selector,
agentMetricTestResult JSON types -- goes with the file.

See bpfman/bpfman-operator#527.
